// src/models/IoK8sKubeAggregatorPkgApisApiregistrationV1APIServiceList/types.ts
import { IoK8sApimachineryPkgApisMetaV1ListMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ListMeta } from '../IoK8sApimachineryPkgApisMetaV1ListMeta';
import { IoK8sKubeAggregatorPkgApisApiregistrationV1APIServiceModelOptions as IoK8sKubeAggregatorPkgApisApiregistrationV1APIService } from '../IoK8sKubeAggregatorPkgApisApiregistrationV1APIService';

export interface IoK8sKubeAggregatorPkgApisApiregistrationV1APIServiceListModelOptions {
  apiVersion?: string;
  items: Array<IoK8sKubeAggregatorPkgApisApiregistrationV1APIService>;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ListMeta;
}